sysutils/u-boot-pinebookpro: Add new port
This is a u-boot slave port for the Pine64 PineBookPro, an RK3399
ARM64 laptop.
The patches have been submitted upstream and will probably be part
of the next u-boot release, but add the port in the meantime in case
anyone wants to start testing/dev FreeBSD on this laptop.
It doesn't have video support for now so one want to use a serial
cable.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@532960 35697150-7ecd-e111-bb59-0022644237b5