Update to 6.0.1.
<ChangeLog>

Upgrade urgency HIGH: This release fixes a crash when builiding against
                      Libc malloc.

Here we revert 8110ba888, an optimization that causes a crash due to a
bug in the code. It does not happen with the default allocator because of
differences between Jemalloc and libc malloc, so this escaped all our
testing but was reported by a user. We'll add back the original optimization
that was reverted here later, after checking what happens: it is not a
critical optimization.

</ChangeLog>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@533927 35697150-7ecd-e111-bb59-0022644237b5