net/dpdk: ix build with 13-CURRENT
The automated package builds for net/dpdk are failing on 13-CURRENT
because of changes to the FreeBSD code in the 13 tree. This update
provides minor tweaks to the code to fix compatibility with BSD 13.

Two issues were observed:
* rename of CPU_NAND to CPU_ANDNOT
* change in vm_page_replace parameters and behaviour

Both issues only required small changes to a single file, and both
fixes are in the diff attached.

PR:		243058
Submitted by:	Bruce Richardson <bruce.richardson@intel.com> (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@535736 35697150-7ecd-e111-bb59-0022644237b5