New port: net/librespeed-cli: Command line client for LibreSeed
This is the CLI part of LibreSpeed which is a very lightweight speedtest.
With the CLI it's possible to run a speedtest from the command line interface.
The server part can be found in www/librespeed.

WWW: https://github.com/librespeed/speedtest-cli

PR:		246609
Submitted by:	Dries Michiels <driesm.michiels@gmail.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@539318 35697150-7ecd-e111-bb59-0022644237b5