vuln.xml: Adjust sqlite version in sqlite entry
Update the sqlite versions affected in the latest sqlite entry.  The entry
failed to take PORTEPOCH into account, and without this fix pkg audit fails
to mark sqlite as vulnerable when it's not updated to the latest version,
since any version with PORTEPOCH set will always be greater than any version
without.

PR:		247149


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@540402 35697150-7ecd-e111-bb59-0022644237b5