www/p5-Web-Machine: create port
Web::Machine provides a RESTful web framework modeled as a state machine.
You define one or more resource classes. Each resource represents a single
RESTful URI end point, such as a user, an email, etc. The resource class can
also be the target for POST requests to create a new user, email, etc.

Each resource is a state machine, and each request for a resource is handled by
running the request through that state machine.

Web::Machine is built on top of Plack, but it handles the full request and
response cycle.

PR:		248041


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543237 35697150-7ecd-e111-bb59-0022644237b5