www/p5-HTTP-Headers-ActionPack: create port
This is a module to handle the inflation and deflation of complex HTTP header
types. In many cases header values are simple strings, but in some cases
they are complex values with a lot of information encoded in them. The goal
of this module is to make the parsing and analysis of these headers as easy as
calling inflate on a compatible object.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543238 35697150-7ecd-e111-bb59-0022644237b5