lang/rust-nightly: Attempt to fix intermittent "can't find crate for `std`" build failures
The location of rustc (found via env::current_exe()) is used to
find the right libstd.  However it might have been "copied" by
creating a hard link to the new location instead.  Like /proc/curproc/file,
KERN_PROC_PATHNAME (used internally by current_exe()) can return
any of the file's multiple paths.  Most of the time it returns the
right rustc path and the build will succeed but occasionally it
will return the "wrong" path and the build fails with:

    error[E0463]: can't find crate for `std`

If this is right a viable workaround should be to never create hard
links during the build, so let's try that.

Also drop the related llvm-config-wrapper workaround.

PR:		248184


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543251 35697150-7ecd-e111-bb59-0022644237b5