devel/ioncube: update to 10.4.1
2020-07-13 	10.4.1 	Fix for parent classes sometimes not being found in the
PHP 7.4 Loader. Fix for incorrect errors arising when a class implemented an
abstract method through a trait in the PHP 7.4 Loader. Corrected build for
Aarch64 PHP 7.4 loaders.

2020-06-29 	10.4.0 	Loaders for PHP 7.4. These will run files produced by
the PHP 7.1 and 7.2 encoders. Fix for an issue with the encoded_paths feature
in Windows.

2020-02-28 	10.4.0 beta 2 	Second beta release of PHP 7.4 Loaders for
Linux. These now run both 7.1- and 7.2-encoded files but do not support IC24.
These beta loaders should not be used in a production environment.

2020-02-27 	10.4.0 beta 	Beta release of PHP 7.4 Loaders for Linux. At
present these only run 7.2-encoded files and should not be used in a production
environment.

2019-10-11 	10.3.9 	Fix for an issue in the PHP 7.3 Loader which could
cause OPcache to be disabled on unencoded files. Fix for an issue in the PHP
7.2 and 7.3 Loaders which could give a spurious warning message if a parameter
of a method with a null value overrides a method with a nullable type.

2019-08-23 	10.3.8 	Fix for a possible crash in the PHP 7.3 Loader when
OPcache was also enabled. Fix for PHP error exceptions not being caught in the
PHP 7 Loaders. Fixes for PHP issues 71980, 75396 and 75786. Fix for an
incorrect call order with certain combinations of foreach, finally and
exception in the PHP 7.1 Loader. Fix for incorrect line numbers being reported
for undefined array index errors. Fix for a potential memory leak with the
encoded paths feature of the Loader. IC24: Minor bug fixes, improvements and
support for future features. Upgrading to the 10.3.8 Loader is particularly
recommended for IC24 users. New 5.4 and 7.2 Loaders custom built for IBM for
ppc64le.

2019-06-25 	10.3.7 	Fix for a potential segmentation fault in the 10.3.5
and 10.3.6 Loaders when including a large file defining an array. The Loader
version now can be obtained using phpversion("ionCube Loader").

2019-06-18 	10.3.6 	Fix for an issue in the PHP 7.3 Loader where a case
within a switch could be triggered even though the variable in the switch was
undefined. Fix for filepaths shown in debug_backtrace and similar functions
being incorrect in some cases in the 10.3.5 Loaders. Fix for possible
segmentation faults in the 10.3.5 Loaders occurring with include file
protection.

2019-06-04 	10.3.5 	The ionCube Loader now has its own section in the
phpinfo() output. That will include IC24 settings where relevant. Fix for some
memory leaks in PHP 5.6-7.3, particularly when including files in a loop. For
undefined variables in assignments, correction to ensure that the line numbers
in error messages are not mangled when line number obfuscation is not in place.
Fix to ensure that the loader will accept large property values that were
accepted by the Encoder. Fix to ensure that external encoding keys set at
runtime using ini_set would be found.

2019-04-07 	10.3.4 	This updates Linux x86 32/64 bit Loders only, and fixes
corruption of the HTTP_HOST server variable if it includes a port number, e.g.
somedomain.com:1081. There is no change for other platforms other than the
version number for consistency.

2019-04-05 	10.3.3 	Fix for callback files not being called from bundled
encoding files.  Added AArch64 (ARM64) as a new standard platform.

2019-01-02 	10.3.2 	Fix for a potential issue with the corruption of
included file paths, particularly on Windows.

PR:		240651
Submitted by:	Svyatoslav Lempert <svyatoslav.lempert@gmail.com> (maintainer)
Reported by:	Eero Hänninen <fax@nohik.ee>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543539 35697150-7ecd-e111-bb59-0022644237b5