rsync: Unbreak and fix depends
rsync now depends on stuff in LOCALBASE. Previously, clang only needed to know
about LOCALBASE if POPT or ICONV was enabled. When those options are off, xxhash
and zstd were not found by configure.

Also, a depend on libssl was missing, and there were some noop reinplaces.

With hat: portmgr


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543637 35697150-7ecd-e111-bb59-0022644237b5