cad/NASTRAN-95: mark broken
Recent changes in gfortran seem to have broken this port so that it doesn't
produce valid results anymore. See [1] for more details.

This port would likely be a good candidate for LLVM Fortran but I currently
don't have time for to try (patches welcome).

Use dos2unix, as it seems necessary but not sufficient for a future fix.
Pet portlint while here.

Approved by:	therry (mentor)

[1] https://github.com/nasa/NASTRAN-95/issues/15


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@543805 35697150-7ecd-e111-bb59-0022644237b5