Add py-forcediphttpsadapter 1.0.1
This module implements a set of requests TransportAdapter, PoolManager,
ConnectionPool and HTTPSConnection with one goal only: to use a specific IP
address when connecting via SSL to a web service without running into SNI
trouble.

The usual technique to force an IP address on an HTTP connection with Requests
is (assuming I want http://example.com/some/path on IP 1.2.3.4):
  requests.get("http://1.2.3.4/some/path", headers={'Host': 'example.com'})

This is useful if I want to specifically test how 1.2.3.4 is responding; for
instance, if example.com is DNS round-robined to several IP addresses and I want
to hit one of them specifically.

WWW: https://github.com/Roadmaster/forcediphttpsadapter


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@546975 35697150-7ecd-e111-bb59-0022644237b5