Fix build with -no-common
While here make the substitution of %%PREFIX%% operational (again?).

It has probably been lost due to make makepatch generating patch files
after the substitution by the actual prefix has been performed in the
post-patch target.

To prevent this issue in the future, the substitution of %%PREFIX%%
by the actual prefix is moved in to the post-configure phase, since the
configure command does not rely on the substitution having been performed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@547529 35697150-7ecd-e111-bb59-0022644237b5