Fix build with -fno-common
While this makes the build process proceed past the point where the it
previously failed with -fno-common, it still fails later in the examples.

There is a comment in a README file that mentions BUILD_EXAMPLES_FROM_TREE,
but I could not find any actual reference to that parameter outside the
README file and the build failed with and without it set.

This commit is meant to document the changes require due to -no-common
and I leave the fix required in lttng-ust-2.9.0/doc/examples/gen-tp for
a later time or another committer.

The PORTREVISION is not bumped, since the build still fails for me.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@547532 35697150-7ecd-e111-bb59-0022644237b5