New port: net-im/zoom: Zoom videoconferencing client
Sound doesn't work yet. It seems that Zoom has some bug and
doesn't show audio devices that the ALSA interface returns
to it. Hopefully this will be fixed by Zoom.

Lack of sound severely limits its usability, but it can still be
useful in these situations:
* A particular Zoom conference isn't supported through the web interface.
  At least video can be watched in such case, vs. nothing at all.
* Video or chat aren't displayed properly through the web interface,
  and this client can be used to augment the experience.

Installing it into /opt because the /opt/zoom path is hardcoded
into the binary.

Zoom didn't define any license or terms of use for their binaries.
It is just a free download with a single click.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@548898 35697150-7ecd-e111-bb59-0022644237b5