Add a build dependency on GNU as
On systems with GNU as in base (up to and including 12.x) there is no need
to provide the GNU binutils from a port.

If the not fully compatible llvm-as should be made available under the name
"as", the condition should be changed to always require binutils from a port
when running on a system after 12.x (which should retain the GNU as in base
for the live-time of that major release).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549608 35697150-7ecd-e111-bb59-0022644237b5