Fix build
The build system includes bsd.sys.mk, which add a -fuse-ld option in a
format that is not accepted by gfortran and thus causes a build failure
of the test programs. This can be worked-around by adding a ".f:" target
to the port's internal Makefile.

While here add a license (PD) and use a relative symbolic link for the
binary's alias name.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549677 35697150-7ecd-e111-bb59-0022644237b5