Fix build
The configure script ran into an infinite loop due to an optimization
that assumes multiplication of 2 positive numbers was always positive
(ignoring wrap-around). Pass "-fwrapv" to the configure script and build
the package with that option.

Two parsers are generated and both use the same global variables. Fix
the build with -fno-common by use of pre-processor macros that rename
the variables used in one of those source files. Passing -p to set a
non-conflicting prefix other than "yy" does not work with implicit
".c.y" rules and was therefore not applicable.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@549724 35697150-7ecd-e111-bb59-0022644237b5