audio/ardour6: Fix broken 32-bit build
This fixes the Intel 32-bit build.
Modification of the original wscript file is no longer necessary.
Setting mmx and 3dnow compiler flags is not really needed.
Setting 3dnow can even be considered as a bug as this is a AMD only extension.
The build_host_supports_sse flag is only used to throw a warning if the set configuration doesn't match with the build machines configuration.
The warning can be ignored.
The build requires a SSE capable CPU (>=i686) which was also the case in previous versions.

PR:		250206
Submitted by:	Michael Beer (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@552098 35697150-7ecd-e111-bb59-0022644237b5