[NEW PORT] www/castor: Graphical browser for text-based inet protocols
Castor is an actively developed graphical browser for text-based protocols
written in Rust that uses the GTK+ toolkit. It currently supports the
Gemini, Gopher and Finger protocols.

Gemini is a modern take on creating a text-based protocol that lives up to
today's standards technically (e.g. it uses TLS). It aims to provide a
"middle ground" between the bare-bones Gopher and the bloated WWW but
leaning towards the former.

PR:		250266
Submitted by:	Michael Reim <kraileth elderlinux org>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@552362 35697150-7ecd-e111-bb59-0022644237b5