New port: database/cutelyst-asql
asql is an asynchronous, C++11-native, Qt5 based SQL Query library.
You can use it to write async queries with natural C++-lambda
and Qt5-slots style code.

Right now there is only a PostgreSQL backend, but others are
planned (and then the port will grow OPTIONS).

This comes from the cutelyst project, and it would conflict with
www/asql is named the "obvious" way, so name it something inbetween.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@553357 35697150-7ecd-e111-bb59-0022644237b5