Add py-mixsimulator 0.2.9.5
MixSimulator is an application with an optimization model for calculating and
simulating the least cost of an energy mix under certain constraints. The
optimizers used are based on the Nevergrad library.

The primary objective of the simulator is to study the relevance of an energy
mix connected to each Inter-connected Grid through the coefficient of usage of
each unit in the production cost.

WWW: https://github.com/Foloso/MixSimulator


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@554737 35697150-7ecd-e111-bb59-0022644237b5