[NEW PORT] devel/py-maturin
Build and publish crates with pyo3, rust-cpython and cffi bindings as well
as Rust binaries as Python packages.

This project is meant as a zero configuration replacement for
setuptools-rust and milksnake.  It supports building wheels for Python 3.5+
on FreeBSD and various other platforms, can upload them to PyPI and has
basic PyPy support.

WWW: https://github.com/PyO3/maturin

PR:		247666
Reviewed by:	tcberner
Differential Revision:	https://reviews.freebsd.org/D26960


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@554832 35697150-7ecd-e111-bb59-0022644237b5