Remove deprecated optional patches
Remove broken options available in the past as additional patches provided
by the project and merged in the latest releases as part of the main source.

time-limit merged in 3.2.3 (6 Aug 2020)
   Added the --stop-after=MINS and --stop-at=DATE_TIME options (with the --time-limit=MINS option
   accepted as an alias for --stop-after). This is an enhanced version of the time-limit patch
   from the patches repo.

atimes merged in 3.2.0 (19 Jun 2020)
   Added the --atimes option based on the long-standing patch (just with some fixes that the patch
   has been needing).

acls merged in 3.2.0 (19 Jun 2020)
   Added the --atimes option based on the long-standing patch (just with some fixes that the patch
   has been needing).

Since those options are not activated by default, and if so, breaks the build,
no PORTREVISION bump is required.

PR:		250998
Submitted by:	 Zoltan ALEXANDERSON BESSE <zab@zltech.eu>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@554934 35697150-7ecd-e111-bb59-0022644237b5