Resign as maintainer of this port of GCC 4.8 which has been unsupported upstream since 2015 and marked deprecated for three years now.
sysutils/memtest86+ and sysutils/uefi-edk2-bhyve are two remaining users
in the entire ports tree which hopefully will be converted at one point.

On the way refer to GCC 10 instead of GCC 9 for a recommended update.

And move the USES block earlier in the Makefile as newer GCC ports do
to honor our coding standards (and pet portlint).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@555437 35697150-7ecd-e111-bb59-0022644237b5