x11/remontoire: add new port
Remontoire is a small (~71Kb) GTK app for presenting keybinding hints
in a compact form suitable for tiling window environments. It is
intended for use with the i3 window manager but it's also able to
display keybindings from any suitably formatted config file.

The program functions by scanning and parsing comments in a specific
format (described directly below), then displaying them in a one-layer
categorized list view. The program stores the state of which sections
are expanded, allowing for use on screens with limited resolution.

https://github.com/regolith-linux/remontoire


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@555670 35697150-7ecd-e111-bb59-0022644237b5