games/gcompris-qt: update to 1.0
To celebrate the 20 years of GCompris, we are pleased to announce the release
of GCompris version 1.0.

This new version contains a new major feature: the addition of the Activity
Settings menu with Dataset selection for more than 50 activities, allowing to
choose more specifically what can be learned on the activities.

Announcement:
	https://gcompris.net/index-en.html#2020-11-19


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@556065 35697150-7ecd-e111-bb59-0022644237b5