emulators/qemu-user-static: fix the build after base r368005
r368005 removed CRIOGET, breaking the qemu-user-static build. We didn't
actually implement it anyways, so this update simply ripped it out rather
than trying to maintain it for old branches.

While here, bring qemu-user-static-devel up to parity with qemu-user-static.
We should be soon approaching the point where we can rebase the former
forward to a newer qemu.

Reported by:	gjb


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@556947 35697150-7ecd-e111-bb59-0022644237b5