Add sysutils/rubygem-choria-mcorpc-support
These are stripped down versions of the legacy Marionette Collective libraries
that are used to facilitate hosting Ruby based MCollective Agents within the
Golang daemon as well as interact with a Choria network from the Ruby API and
mco CLI

It's not possible to run the old MCollective daemon using this code or the
resulting gem.

This fork was made off version 2.12.0 of The Marionette Collective and have
diverged since then, the changes made are not compatible with legacy
MCollective.

WWW: http://rubygems.org/gems/choria-mcorpc-support

With hat:	puppet


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@557099 35697150-7ecd-e111-bb59-0022644237b5