Actually update the newly created port lang/mono6.8 to 6.8.
This port is a WIP.  It compiles and packages on both amd64 and powerpc64.
However, the ports depending on it do *not*.  I will continue working on
this.  Because of this, assign maintainership to myself in the meantime.

This corresponds to the most recent patches in D23300 except with fixes
to the OPTIONS handling by myself.

Submitted by:		prj_rootwyrm.com
Differential Revision:	D23300


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@557880 35697150-7ecd-e111-bb59-0022644237b5