Update KDE Frameworks to 5.77.0
December 12, 2020. KDE today announces the release of KDE Frameworks 5.77.0.

KDE Frameworks are 83 addon libraries to Qt which provide a wide variety of
commonly needed functionality in mature, peer reviewed and well tested
libraries with friendly licensing terms. For an introduction see the KDE
Frameworks release announcement.

This release is part of a series of planned monthly releases making
improvements available to developers in a quick and predictable manner.

Announcement:
	https://kde.org/announcements/kde-frameworks-5.77.0/

PR:		251792
Exp-run by:	antoine


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@558298 35697150-7ecd-e111-bb59-0022644237b5