Add py-capturer 3.0
The capturer package makes it easy to capture the stdout and stderr streams of
the current process and subprocesses. Output can be relayed to the terminal in
real time but is also available to the Python program for additional processing.

WWW: https://capturer.readthedocs.io/en/latest/
WWW: https://github.com/xolox/python-capturer


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@558472 35697150-7ecd-e111-bb59-0022644237b5