Update net-im/neochat to its first public release
NeoChat is a client for the Matrix instant-messaging network;
NeoChat is a convergent client available on desktop and mobile alike.

This is the first public release of NeoChat (previous ports were of
development snapshots). A new dependency has appeared, to allow
editing of screenshots and photos before sending them. Release
announcement is at
	https://carlschwan.eu/2020/12/23/announcing-neochat-1.0-the-kde-matrix-client/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@560251 35697150-7ecd-e111-bb59-0022644237b5