Follow up on graphics/py-pillow6 port removal in r559976
The graphics/py-pillow6 port was removed in r559976 as it was for
Python 2.7 only and no longer referenced in the default cases.
However it was still defined in Mk/Uses/python.mk in PY_PILLOW,
leaving a dangling reference.

Remove it from Mk/Uses/python.mk, leaving only the Python 3
version of that port, graphics/py-pillow, defined and bump the
minimal verion of Python to 3.6 for affected ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@560425 35697150-7ecd-e111-bb59-0022644237b5