Follow up on devel/py-futures port removal in r559976
The devel/py-futures port was removed in r559976 as it was for
Python 2.7 only and no longer referenced in the default cases.
However it was still defined in Mk/Uses/python.mk in PY_FUTURES,
leaving a dangling reference.

Remove it from Mk/Uses/python.mk and from affected ports, as
it was already an empty definition for Python 3.6+

Bump minimal Python version of security/theonionbox to 3.6+


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@560434 35697150-7ecd-e111-bb59-0022644237b5