Add py-cwcwidth 0.1.1
cwcwidth provides Python bindings for wcwidth and wcswidth functions defined in
POSIX.1-2001 and POSIX.1-2008 based on Cython. These functions compute the
printable length of a unicode character/string on a terminal. The module
provides the same functions as wcwidth and its behavior is compatible.

On systems not conforming to POSIX.1-2001 and POSIX.1-2008, Markus Kuhn's
implementation is used to provide the functionality.

WWW: https://github.com/sebastinas/cwcwidth


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@562718 35697150-7ecd-e111-bb59-0022644237b5