Fix build in poudriere
Fetching a cmake helper file was (rightfully) blocked when building with
poudriere, leding to an abort in the configure phase.

Instead of fetching the cmake helper as an additional distfile, provide
the relevant functionality in a local version that uses port system
information instead of trying to obtain the version number from the
sources.

Reported by:	kai


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@564105 35697150-7ecd-e111-bb59-0022644237b5