Update to 2.2.0
With the last release, πmenu now uses XRender to draw anti-aliased graphics
(the line separators, and the slices).

Also, you do not need xclickroot anymore: with the -r 3 option, pie menu
can be run in background (by following it with a &) and it will watch for
clicks on the root window. I included this xclickroot functionality into
πmenu so it is always loaded, and it does not have the overhead of
initiating and drawing stuff.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@564111 35697150-7ecd-e111-bb59-0022644237b5