Add a new port editors/marker
Marker is a simple yet robust markdown editor.

The primary goal of Marker is to provide a powerful document editing
experience, without making assumptions about the workflow of the
user. Marker's base configuration is very simple, but it can be
customized to meet the needs of any workflow!

WWW: https://fabiocolacio.github.io/Marker/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@564832 35697150-7ecd-e111-bb59-0022644237b5