Add KDE_INVENT convenience for KDE ports
KDE has its own GitLab instance. This behaves slightly different
from other GitLab -- there's not really a "GL_ACCOUNT" involved,
although the repositories are filed under various categories
that behave like accounts do on GitLab.com.

Some -- not many -- ports of KDE software pick up unreleased
software, or software in-between releases. Those tend to use
the GitHub mirror. That mirror should not be considered a good
source for the software; it's better to stick close to the real source.

With KDE_INVENT it is immediately clear that a port comes from KDE
sources, but not released work. It is a shortcut for setting GL_
values for site, account (category) etc. While there's only a handful
of ports now that would need it (and we would prefer to **not**
package unreleased software) the "signal value" is a good one.

Reviewed by:	tcberner


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@565178 35697150-7ecd-e111-bb59-0022644237b5