net/wireguard: Update to 1.0.20210315 and drop BASHCOMPLETION option
BASHCOMOPLETION option was dropped but we install shell completion files
unconditionally to follow porters handbook.

This is the first wireguard-tools version that supports FreeBSD's kernel
implementation. It is compatible with if_wg in FreeBSD 14-CURRENT which
will likely be backported in the future but incompatible with the if_wg
code that was present in stable/13 up to 13.0-RC3.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@568486 35697150-7ecd-e111-bb59-0022644237b5