Update deskutils/semantik to latest upstream release
This is prompted by the imminent release of KDE Frameworks 5.80,
which had fallout with this software. I have no patience with
waf as a build system, so this adds some hacks to include
absolute paths to the KF5 includes, rather than -I../../../KConfig
-I../../../../KConfig (on the off chance you might be one source
level higher or lower; this is what the build system is doing
without hacks).

PR:		254256 (fallout)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@568764 35697150-7ecd-e111-bb59-0022644237b5