devel/gettext-po-mode: New port to install po-mode files for emacs
-BEGIN pkg-descr---------------------------
GNU gettext is a framework of libraries and tools for internationalisation
and localisation of software.

This package contains the po-mode for the emacs editor.

WWW: https://www.gnu.org/software/gettext/
-END pkg-descr-----------------------------

This cannot be part of gettext-tools because it would create a
dependency loop through emacs (which we use to byte-compile).

In order to override the MAINTAINER, modify gettext/Makefile.common
to use MAINTAINER?= to accomodate overrides.

Next step: we should modularize up the emacs site-start.el.
Initial email sent to emacs@.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@568789 35697150-7ecd-e111-bb59-0022644237b5